Fix: Wrong Memory Access Protocol.
Atom earlier used a wrong memory access protocol whereby a
word/half-word/byte access can be made at any address and there was
no constaint for alignment. This however never resulted in any issues
because the memory was simulated in such a way in Cpp (AtomBones).
This is fixed in this commit.

passing riscv-arch-tests

Other changes:
- Add gtkwv save files.
- default trace file using --trace CLI option : build/trace/trace.vcd
- add more scar tests
- remove scar test numberings
- separate linker script for scar